Adjust naming convention for SmallVisualiserWidget.cs
